+++ R/plot_ID166_pdf.R
@@ -11,14 +11,15 @@
 #' @param ylabels Logical. Draw y-axis labels.
 #' @param ylim Optional y-axis limits.
 #' @param base_size Numeric. Base font size in points.
-#' @param title_cex Numeric. Multiplier for the plot title size.
+#' @param plot_title_cex Numeric. Multiplier for the plot title size.
 #' @param count_label_cex Numeric. Multiplier for per-class count labels.
 #' @param block_label_cex Numeric. Multiplier for category block labels.
 #' @param class_label_cex Numeric. Multiplier for major class labels.
-#' @param x_label_cex Numeric. Multiplier for x-axis labels.
+#' @param axis_text_x_cex Numeric. Multiplier for x-axis labels.
 #' @param bottom_label_cex Numeric. Multiplier for bottom labels.
-#' @param axis_title_cex Numeric. Multiplier for the y-axis title size.
-#' @param axis_text_cex Numeric. Multiplier for the y-axis tick label size.
+#' @param axis_title_x_cex Numeric. Multiplier for the x-axis title size. Currently has no effect in this function.
+#' @param axis_title_y_cex Numeric. Multiplier for the y-axis title size.
+#' @param axis_text_y_cex Numeric. Multiplier for the y-axis tick label size.
 #' @param show_counts Logical or NULL. Auto-detect if NULL.
 #'
 #' @return Invisibly returns `NULL`.
@@ -44,14 +45,15 @@ plot_ID166_pdf <- function(
   ylabels = TRUE,
   ylim = NULL,
   base_size = 11,
-  title_cex = 0.8,
+  plot_title_cex = 0.8,
   count_label_cex = 0.6,
   block_label_cex = 0.65,
   class_label_cex = 0.8,
-  x_label_cex = 0.5,
+  axis_text_x_cex = 0.5,
   bottom_label_cex = 0.65,
-  axis_title_cex = 1.0,
-  axis_text_cex = 0.8,
+  axis_title_x_cex = 1.0,
+  axis_title_y_cex = 1.0,
+  axis_text_y_cex = 0.8,
   show_counts = NULL
 ) {
   stopifnot(nrow(catalog) == 166)
@@ -76,14 +78,15 @@ plot_ID166_pdf <- function(
         ylabels = ylabels,
         ylim = ylim,
         base_size = base_size,
-        title_cex = title_cex,
+        plot_title_cex = plot_title_cex,
         count_label_cex = count_label_cex,
         block_label_cex = block_label_cex,
         class_label_cex = class_label_cex,
-        x_label_cex = x_label_cex,
+        axis_text_x_cex = axis_text_x_cex,
         bottom_label_cex = bottom_label_cex,
-        axis_title_cex = axis_title_cex,
-        axis_text_cex = axis_text_cex,
+        axis_title_x_cex = axis_title_x_cex,
+        axis_title_y_cex = axis_title_y_cex,
+        axis_text_y_cex = axis_text_y_cex,
         show_counts = show_counts
       )
       page_plots[[length(page_plots) + 1]] <- p